move cert replacement  test
